Fix/reload rendering (#7)
* Fix OptionsStep's problem when loaded from the cache

OptionsStep when loaded from the cache had a problem, where the chosen option could not be matched up with either of the Options that was loaded from the cache. This matching was done using `filter()` where `===` operator is used. This caused problem as `===` operator only checks for Object reference equality. This has been fixed with the `deep-copy` package where Object's inner value is compared.

* Fix Chat reload problem: the textmessages continues even after reloading

When `renderedSteps` are loaded from `localStorage` their
`triggerNextStep()` are blocked by specifying
`renderedStep.rendered = true`. This behavior should not happen on the
last step, as last step needs to continue rendering upcoming step.

* Fix chat reload problem: when reloading at UserStep

* Reassign validator & parser to current step when loaded from cache

When a current step which is a user step is loaded from cache, it loses all reference to its validator and parser function. This occurs as `stringify()` does not stringify functions. Thus, the `validator` and `parser` function is reassigned to user step from the provided `steps`. In not done so, it would reject some of the perfectly valid user input.

* Fix bug for reloading at UserStep and Add tests for same at Option

Bug regarding reloading at `UserStep` is that, when the chat is reloaded, and the steps are extracted from the `localStorage` cache, `parser` and `validator` is not assigned to the `currentStep` which is a `UserStep`.

Same bug resided when Chat is reloaded at `OptionStep`. But, that has already been solved by earlier commit. This commit just adds tests for it.

* Allow chat to proceed when reloaded at a TextStep

* Better fix for storage

* Fix chat stopping problem when reloaded at TextStep

* Add tests for storage.js

To ensure proper storage and parsing of chat state using localstorage. This checks for several edge cases where the chat flow might stop or fail due to incorrect storage and parsing.